Throw PythonException for ANTLR syntax errors
This catches parser syntax errors and throws a C++ PythonException while
also avoiding trying to translate a tree with syntax errors.

The Python exception `SyntaxError` (normally meaning a Python code
syntax error) is raised out of the Python extension code and repurposed
to represent a syntax error in the target langauge and to differentiate
from the `RuntimeError` and other potential Python exception types
from other areas. The use of `SyntaxError` could cause
confusion if the callee catches the `SyntaxError` and there actually is
a syntax error in one of the imported ANTLR modules, but this is
unlikely.