Add ANTLR list label support
This adds support for ANTLR "list labels" which are are a list of tokens
or context objects generated from grammar containing `+=` operators.

Previously speedy-antlr would generate C++ code that referenced
nonexistent context member names beginning with "_" because it extracted
the same-named attributes from the Python parser that were generated by
ANTLR for rules referenced by list labels.

TODO: `validate.py` does not check list labels. I found during testing
that it also does not seem to handle other items in more complicated
grammars than speedy-antlr-example, so is left for a future commit. For
instance, contexts with no children are currently an issue. The Python
parser sets children to `None` in this case and validate calls `len()`
on it. Fixing this issue leads to validate properly failing because
start/stop is not translated from C++ for nodes that have no children,
although the Python parser does set start/stop in this case. So whether
to ignore in validate or fix in the translator is TBD.