Automate and assimilate speedy-antlr-example
Main addition is the package.py tool that is a first pass at automating
the application of speedy-antlr-example to a given grammar/project.
The above uses template project files that include the ANTLR C++
runtime source.

A good improvement would be to automate the fetching
of the ANTLR code depending on the version of ANTLR tool being
used or at least add an option to point to the code that the user
downloaded manually. Then the ANTLR code could be removed here.